+++ src/main.rs
@@ -3,6 +3,7 @@ use std::env::var;
 use clap::{App, SubCommand};
 
 mod lib {
+    pub mod brew;
     pub mod cargo;
     pub mod env;
     pub mod favourites;